requirements.txt
Django==1.9.2
djangorestframework==3.3.2
enum34==1.1.2